requirements.txt
asyncio==3.4.3
aiohttp==3.4.4
aioredis==1.1.0
gino==0.7.5
python-socketio==2.0.0
sanic==0.7.0
